requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.14.0
scikit-learn>=1.2.0
streamlit
requests
python-dotenv
